+++ masonry.js
@@ -18,7 +18,7 @@
 /**
  * Masonry
  */
-(function (window, factory) {
+((window, factory) => {
     'use strict';
 
     if (typeof define === 'function' && define.amd) {
@@ -40,7 +40,7 @@
             window.getSize
         );
     }
-}(window, function factory (Outlayer, getSize) {
+})(window, (Outlayer, getSize) => {
     'use strict';
 
     // create an Outlayer layout class
@@ -264,4 +264,4 @@
     };
 
     return Masonry;
-}));
+});